Add spread regress-stack distro job.
A GitHub matrix is populated by querying the Launchpad API for
supported Ubuntu versions.  This relieves us of the perpetual
mainteneance task of adding new versions to the matrix.

Test preparation and execution steps are expressed in spread YAML
which allows parallel execution of the same test code in multiple
environments.

The spread systems list is generated by running the `boot.sh` shell
script.

When running in GitHub Actions systems are generated on the back of
input from the GitHub matrix.

For local execution the `boot.sh` script can reach out to Launchpad
API directly to build an up to date list of Ubuntu versions.

For each Ubuntu version three systems are allocated to accomodate
for testing with different states of system packaging:
* Default APT pockets.
* Add -proposed APT pocket with preference to install.
* Add security-proposed PPA.

A future improvement might be to make use of spread features to
quickly iterate over variations of tests on the same system.

The current challenges preventing us from doing it now is:
* The nature of the test leaves large amounts of crud in the system,
  which is difficult to fully clean up between variations.
* Variations involve changes to the system such as APT sources and
  consequently package upgrades.  Consequently the two approaches
  in reality tests different behaviors, i.e. fresh install vs.
  upgrade.

Output of test execution is collected by spread and made available
on GitHub Actions summary page.

Signed-off-by: Frode Nordahl <fnordahl@ubuntu.com>