Redirect stderr to stdout
When running regress-stack the stdout is capture in prepare-output.txt
and execute-output, although the stderr is left out, this prevents from
getting  a full copy of it since spread will discard parts of it when
too much of it is being written, traces of the log below will be found
in those scenarios.

WARNING: spread-1-ubuntu-plucky-ppa (lxd:ubuntu-plucky-ppa:tests/spread/regress-stack) running late. Current output:
-----
(... 1005 lines above ...)

Signed-off-by: Felipe Reyes <felipe.reyes@canonical.com>